+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* Improve visual design for marking up farmlands and resource income
+
 ## [0.8.1] - 2023-10-14
 
 ### Fixed